feat: migrate components in api explorer (#3494)
* migrate api explorer

* fix style issues

* add missing dep

* separate styling

* fix imports

* fix tree navigation

* fix actions

* remove unnecessary martin top

* update styles

* update lock file

* fix link styles

* fix alignment

* fix build

* revert lock file changes

* revert package changes

* remove stackshim

* fix tests

* fix linting
