+++ src/app/views/sidebar/resource-explorer/ResourceLink.tsx
@@ -1,8 +1,6 @@
-import {
-  getId, getTheme, IconButton, INavLink,
-  ITooltipHostStyles, mergeStyleSets, TooltipHost
-} from '@fluentui/react';
-import { CSSProperties, useEffect } from 'react';
+import { Tooltip, Button, Badge } from '@fluentui/react-components'
+import { SubtractSquare20Regular, AddSquare20Regular, DocumentText20Regular } from '@fluentui/react-icons';
+import { useEffect } from 'react';
 
 import { useAppSelector } from '../../../../store';
 import { componentNames, eventTypes, telemetry } from '../../../../telemetry';
@@ -10,14 +8,15 @@ import { IResourceLink, ResourceOptions } from '../../../../types/resources';
 import { validateExternalLink } from '../../../utils/external-link-validation';
 import { translateMessage } from '../../../utils/translate-messages';
 import { existsInCollection, setExisting } from './resourcelink.utils';
-import { getStyleFor } from '../../../utils/http-methods.utils';
+import { useStyles } from './resourceLinkStyles';
 
 interface IResourceLinkProps {
-  link: INavLink;
+  link: IResourceLink;
   resourceOptionSelected: Function;
-  classes: any;
   version: string;
 }
+
+type Colors = 'brand' | 'danger' | 'important' | 'informative' | 'severe' | 'subtle' | 'success' | 'warning'
 
 const ResourceLink = (props: IResourceLinkProps) => {
   const { version } = props;
@@ -30,58 +29,15 @@ const ResourceLink = (props: IResourceLinkProps) => {
     setExisting(resourceLink, existsInCollection(link, paths, version));
   }, [paths])
 
-  const showButtons = {
-    div: {
-      visibility: 'visible'
-    }
-  };
+  const colors: Record<string, Colors> = {
+    'GET': 'brand',
+    'POST': 'success',
+    'PATCH': 'severe',
+    'DELETE': 'danger',
+    'PUT': 'warning'
+  }
 
-  const linkStyle = mergeStyleSets(
-    {
-      link: {
-        display: 'flex', lineHeight: 'normal', width: '100%', overflow: 'hidden', justifyContent: 'space-between',
-        div: {
-          visibility: 'hidden',
-          overflow: 'hidden',
-          marginTop: 2,
-          marginLeft: 'auto'
-        },
-        selectors: {
-          ':hover': { background: getTheme().palette.neutralLight, ...showButtons },
-          ':focus-within': showButtons,
-          '.is-selected &': showButtons
-        }
-      },
-      resourceLinkNameContainer: {
-        textAlign: 'left', flex: '1', overflow: 'hidden', display: 'flex', marginTop: '4px', paddingLeft: '4px'
-      },
-      resourceLinkText: { textOverflow: 'ellipsis', overflow: 'hidden', whiteSpace: 'nowrap', marginTop: '6px' }
-    }
-  );
-
-  const iconButtonStyles = {
-    root: { marginRight: 1 },
-    menuIcon: { fontSize: 16, padding: 5 }
-  };
-
-  const methodButtonStyles: CSSProperties = {
-    background: getStyleFor(resourceLink.method!),
-    textTransform: 'uppercase',
-    padding: '4px',
-    width: '50px',
-    color: 'white',
-    alignSelf: 'center',
-    font: 'bold 12px/16px "Segoe UI", "Segoe WP", "Helvetica Neue", "Nimbus Sans L", Arial, sans-serif',
-    textAlign: 'center',
-    marginBottom: '4px'
-  };
-
-  const tooltipId = getId('tooltip');
-  const buttonId = getId('targetButton');
-  const documentButton = getId('documentButton');
-  const documentButtonTooltip = getId('documentButtonTooltip');
-  const removeCollectionButton = getId('removeCollectionButton');
-  const removeCollectionButtonTooltip = getId('removeCollectionButtonTooltip');
+  const linkStyles = useStyles();
 
   const openDocumentationLink = () => {
     window.open(resourceLink.docLink, '_blank');
@@ -100,95 +56,139 @@ const ResourceLink = (props: IResourceLinkProps) => {
     validateExternalLink(documentationLink || '', componentNames.AUTOCOMPLETE_DOCUMENTATION_LINK, documentationLink);
   }
 
-  const calloutProps = { gapSpace: 0 };
-  const hostStyles: Partial<ITooltipHostStyles> = { root: { display: 'inline-block' } };
-
   setExisting(resourceLink, existsInCollection(link, paths, version));
 
-  const handleAddToCollectionClick = (event: any) => {
+  const handleAddToCollectionClick = (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault();
     event.stopPropagation();
     props.resourceOptionSelected(ResourceOptions.ADD_TO_COLLECTION, link);
   }
 
-  const handleRemoveFromCollectionClick = (event: any) => {
+  const handleRemoveFromCollectionClick = (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault();
     event.stopPropagation();
     props.resourceOptionSelected(ResourceOptions.REMOVE_FROM_COLLECTION, link);
   }
 
-  return <span className={linkStyle.link} tabIndex={0}>
-    {resourceLink.method ?
-      <span className={linkStyle.resourceLinkNameContainer}>
-        <span style={methodButtonStyles}>
-          {resourceLink.method}
-        </span>
-      </span>
-      :
-      <span className={linkStyle.resourceLinkNameContainer}>
-        <span className={linkStyle.resourceLinkText}>
-          {resourceLink.name}
-        </span>
-      </span>
-    }
-
-    <div>
-      {resourceLink.isInCollection ? <TooltipHost
-        content={translateMessage('Remove from collection')}
-        id={removeCollectionButtonTooltip}
-        calloutProps={calloutProps}
-        styles={hostStyles}
-      >
-        <IconButton
-          ariaLabel={translateMessage('Remove from collection')}
-          role='button'
-          id={removeCollectionButton}
-          aria-describedby={removeCollectionButtonTooltip}
-          styles={iconButtonStyles}
-          menuIconProps={{ iconName: 'BoxSubtractSolid' }}
-          onClick={handleRemoveFromCollectionClick}
+  return (
+    <div className={linkStyles.link}>
+      <ResourceLinkNameContainer resourceLink={resourceLink} linkStyles={linkStyles} colors={colors} />
+      {resourceLink.method && (
+        <ResourceLinkActions
+          resourceLink={resourceLink}
+          iconButtonStyles={linkStyles}
+          openDocumentationLink={openDocumentationLink}
+          handleAddToCollectionClick={handleAddToCollectionClick}
+          handleRemoveFromCollectionClick={handleRemoveFromCollectionClick}
         />
-      </TooltipHost> :
-        <TooltipHost
-          content={translateMessage('Add to collection')}
-          id={tooltipId}
-          calloutProps={calloutProps}
-          styles={hostStyles}
-        >
-          <IconButton
-            ariaLabel={translateMessage('Add to collection')}
-            role='button'
-            id={buttonId}
-            aria-describedby={tooltipId}
-            styles={iconButtonStyles}
-            menuIconProps={{ iconName: 'BoxAdditionSolid' }}
-            onClick={handleAddToCollectionClick}
-          />
-        </TooltipHost>}
-
-      {resourceLink.method &&
-        <TooltipHost
-          content={resourceLink.docLink ? translateMessage('Read documentation')
-            : translateMessage('Query documentation not found')}
-          id={documentButtonTooltip}
-          calloutProps={{ gapSpace: 0, target: `#${documentButton}` }}
-          styles={hostStyles}
-        >
-          <IconButton
-            aria-label={translateMessage('Read documentation')}
-            role='button'
-            id={documentButton}
-            disabled={!resourceLink.docLink}
-            aria-describedby={documentButtonTooltip}
-            styles={iconButtonStyles}
-            onClick={() => openDocumentationLink()}
-            menuIconProps={{ iconName: 'TextDocument' }}
-          />
-        </TooltipHost>
-      }
+      )}
     </div>
-    &nbsp;
-  </span>
+  );
 }
 
+const ResourceLinkNameContainer = ({
+  resourceLink,
+  linkStyles,
+  colors
+}: {
+  resourceLink: IResourceLink,
+  linkStyles: any,
+  colors: Record<string, Colors>
+}) => (
+  resourceLink.method ? (
+    <span className={linkStyles.resourceLinkNameContainer}>
+      <Badge
+        className={linkStyles.badge}
+        size='medium'
+        color={colors[resourceLink.method]}
+        aria-label={'http method ' + resourceLink.method + ' for'}>
+        {resourceLink.method}
+      </Badge>
+    </span>
+  ) : (
+    <span className={linkStyles.resourceLinkNameContainer}>
+      <span className={linkStyles.resourceLinkText}>
+        {resourceLink.name}
+      </span>
+    </span>
+  )
+);
+
+const ResourceLinkActions = ({
+  resourceLink,
+  iconButtonStyles,
+  openDocumentationLink,
+  handleAddToCollectionClick,
+  handleRemoveFromCollectionClick
+}: {
+  resourceLink: IResourceLink,
+  iconButtonStyles: any,
+  openDocumentationLink: () => void,
+  handleAddToCollectionClick: (event: React.MouseEvent<HTMLButtonElement>) => void,
+  handleRemoveFromCollectionClick: (event: React.MouseEvent<HTMLButtonElement>) => void
+}) => (
+  <div className={`${iconButtonStyles.actions} actions`}>
+    {resourceLink.method && (
+      <Tooltip
+        withArrow
+        content={
+          resourceLink.docLink? translateMessage('Read documentation')
+            : translateMessage('Query documentation not found')
+        }
+        relationship='label'
+      >
+        {resourceLink.docLink ? (
+          <Button
+            aria-label={translateMessage('Read documentation')}
+            id='documentButton'
+            aria-disabled={!resourceLink.docLink}
+            className={iconButtonStyles.root}
+            icon={<DocumentText20Regular />}
+            appearance='transparent'
+            onClick={() => openDocumentationLink()}
+          /> ) :
+          <Button
+            disabled
+            aria-label={translateMessage('Read documentation')}
+            id='documentButton'
+            aria-disabled
+            appearance='transparent'
+            className={iconButtonStyles.root}
+            icon={<DocumentText20Regular />}
+          />}
+      </Tooltip>
+    )}
+    {resourceLink.isInCollection ? (
+      <Tooltip
+        withArrow
+        content={translateMessage('Remove from collection')}
+        relationship='label'>
+        <Button
+          aria-label={translateMessage('Remove from collection')}
+          id='removeCollectionButton'
+          appearance='transparent'
+          className={iconButtonStyles.root}
+          icon={<SubtractSquare20Regular />}
+          onClick={handleRemoveFromCollectionClick}
+        />
+      </Tooltip>
+    ) : (
+      <Tooltip
+        withArrow
+        content={translateMessage('Add to collection')}
+        relationship='label'
+      >
+        <Button
+          aria-label={translateMessage('Add to collection')}
+          id='targetButton'
+          appearance='transparent'
+          aria-describedby='tooltip'
+          className={iconButtonStyles.root}
+          icon={<AddSquare20Regular />}
+          onClick={handleAddToCollectionClick}
+        />
+      </Tooltip>
+    )}
+  </div>
+);
 export default ResourceLink;
