+++ src/app/views/sidebar/resource-explorer/resource-explorer.utils.ts
@@ -1,5 +1,3 @@
-import { INavLinkGroup } from '@fluentui/react';
-
 import {
   IResource, IResourceLabel, IResourceLink, Method, ResourceLinkType, ResourceMethod, ResourcePath
 } from '../../../../types/resources';
@@ -15,7 +13,7 @@ export function createResourcesList(
   source: IResource[],
   version: string,
   searchText?: string
-): INavLinkGroup[] {
+): IResourceLink[] {
   function getLinkType({ segment, links }: any): ResourceLinkType {
     const isGraphFunction = segment.startsWith('microsoft.graph');
     const hasChildren = links && links.length > 0;
@@ -54,8 +52,8 @@ export function createResourcesList(
     const type = getLinkType({ ...info, links: versionedChildren });
     const enclosedCounter =
       versionedChildren && versionedChildren.length > 0
-        ? ` (${versionedChildren.length})`
-        : '';
+        ? versionedChildren.length
+        : null;
 
     // if segment name does not contain search text, then found text is in child, so expand this link
     const isExpanded =
@@ -70,7 +68,7 @@ export function createResourcesList(
     return {
       key,
       url: key,
-      name: `${segment}${enclosedCounter}`,
+      name: segment,
       labels,
       isExpanded,
       parent,
@@ -79,7 +77,8 @@ export function createResourcesList(
       method: method?.toUpperCase(),
       type,
       links: versionedChildren,
-      docLink: docLink ? docLink : getLink(labels, version, method)
+      docLink: docLink ? docLink : getLink(labels, version, method),
+      count: enclosedCounter
     };
   }
 
@@ -133,11 +132,7 @@ export function createResourcesList(
     ''
   );
 
-  return [
-    {
-      links: navLink.links
-    }
-  ];
+  return navLink.links;
 }
 
 export function generateKey(method: string | undefined, paths: string[], version: string) {
@@ -171,24 +166,21 @@ export function getCurrentTree({
   level,
   resourceItems,
   version
-}: ITreeFilter): INavLinkGroup {
-  let currentTree = createResourcesList(resourceItems, version)[0];
+}: ITreeFilter): IResourceLink {
+  const currentTree = createResourcesList(resourceItems, version)[0];
   const filters = paths.slice(1, level + 1);
+  const currentPaths = ['/'];
   filters.forEach((key: string) => {
-    const linkedKey = findLinkByName(currentTree, key);
-    if (linkedKey) {
-      currentTree = linkedKey;
+    if (key) {
+      currentPaths.push(key);
+    } else {
+      throw new Error(`Path segment "${key}" not found in the current tree.`);
     }
   });
+
+  currentTree.paths = currentPaths;
   return currentTree;
-}
 
-function findLinkByName(list: any, filter: string): INavLinkGroup {
-  return list.links.find((k: any) => removeCounter(k.name) === filter);
-}
-
-export function removeCounter(title: string): string {
-  return title.split(' (')[0].trim();
 }
 
 export function getAvailableMethods(
